Rename GenerateOperationIdUseCase to OperationId
Move from domain/usecase to domain/model as it's a utility, not a use case.

- Rename to OperationId with generate() method
- Update usage in OfflineFirstMemosRepository
- Add tests for OperationId
