+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/sync/domain/model/OperationId.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.sync.domain.usecase
+package space.be1ski.vibits.shared.feature.sync.domain.model
 
 import kotlin.random.Random
 
@@ -6,10 +6,10 @@ private const val RANDOM_MIN = 10000
 private const val RANDOM_MAX = 99999
 
 /**
- * Generates a unique ID for sync operations.
+ * Utility for generating unique sync operation IDs.
  */
-object GenerateOperationIdUseCase {
-  operator fun invoke(): String {
+object OperationId {
+  fun generate(): String {
     val timestamp =
       kotlin.time.Clock.System
         .now()
